+++ ndk/src/media/image_reader.rs
@@ -6,9 +6,9 @@
 
 use crate::media_error::{construct, construct_never_null, MediaError, MediaStatus, Result};
 use crate::native_window::NativeWindow;
+use crate::utils::abort_on_panic;
 use num_enum::{IntoPrimitive, TryFromPrimitive};
 use std::{
-    convert::TryInto,
     ffi::c_void,
     fmt::{self, Debug, Formatter},
     mem::MaybeUninit,
@@ -16,7 +16,8 @@ use std::{
 };
 
 #[cfg(feature = "api-level-26")]
-use std::os::unix::io::RawFd;
+// TODO: Import from std::os::fd::{} since Rust 1.66
+use std::os::unix::io::{FromRawFd, IntoRawFd, OwnedFd};
 
 #[cfg(feature = "api-level-26")]
 use crate::hardware_buffer::{HardwareBuffer, HardwareBufferUsage};
@@ -60,7 +61,7 @@ pub struct ImageReader {
 }
 
 impl Debug for ImageReader {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         f.debug_struct("ImageReader")
             .field("inner", &self.inner)
             .field(
@@ -128,10 +129,12 @@ impl ImageReader {
             context: *mut c_void,
             reader: *mut ffi::AImageReader,
         ) {
-            let reader = ImageReader::from_ptr(NonNull::new_unchecked(reader));
-            let listener: *mut ImageListener = context as *mut _;
-            (*listener)(&reader);
-            std::mem::forget(reader);
+            abort_on_panic(|| {
+                let reader = ImageReader::from_ptr(NonNull::new_unchecked(reader));
+                let listener: *mut ImageListener = context as *mut _;
+                (*listener)(&reader);
+                std::mem::forget(reader);
+            })
         }
 
         let mut listener = ffi::AImageReader_ImageListener {
@@ -154,11 +157,13 @@ impl ImageReader {
             reader: *mut ffi::AImageReader,
             buffer: *mut ffi::AHardwareBuffer,
         ) {
-            let reader = ImageReader::from_ptr(NonNull::new_unchecked(reader));
-            let buffer = HardwareBuffer::from_ptr(NonNull::new_unchecked(buffer));
-            let listener: *mut BufferRemovedListener = context as *mut _;
-            (*listener)(&reader, &buffer);
-            std::mem::forget(reader);
+            abort_on_panic(|| {
+                let reader = ImageReader::from_ptr(NonNull::new_unchecked(reader));
+                let buffer = HardwareBuffer::from_ptr(NonNull::new_unchecked(buffer));
+                let listener: *mut BufferRemovedListener = context as *mut _;
+                (*listener)(&reader, &buffer);
+                std::mem::forget(reader);
+            })
         }
 
         let mut listener = ffi::AImageReader_BufferRemovedListener {
@@ -170,10 +175,13 @@ impl ImageReader {
         MediaError::from_status(status)
     }
 
+    /// Get a [`NativeWindow`] that can be used to produce [`Image`]s for this [`ImageReader`].
+    ///
+    /// <https://developer.android.com/ndk/reference/group/media#aimagereader_getwindow>
     pub fn get_window(&self) -> Result<NativeWindow> {
         unsafe {
             let ptr = construct_never_null(|res| ffi::AImageReader_getWindow(self.as_ptr(), res))?;
-            Ok(NativeWindow::from_ptr(ptr))
+            Ok(NativeWindow::clone_from_ptr(ptr))
         }
     }
 
@@ -206,11 +214,15 @@ impl ImageReader {
         }
     }
 
+    /// Acquire the next [`Image`] from the image reader's queue asynchronously.
+    ///
     /// # Safety
-    /// If the returned file descriptor is not `None`, it must be awaited before attempting to access the Image returned.
+    /// If the returned file descriptor is not [`None`], it must be awaited before attempting to
+    /// access the [`Image`] returned.
+    ///
     /// <https://developer.android.com/ndk/reference/group/media#aimagereader_acquirenextimageasync>
     #[cfg(feature = "api-level-26")]
-    pub unsafe fn acquire_next_image_async(&self) -> Result<(Image, Option<RawFd>)> {
+    pub unsafe fn acquire_next_image_async(&self) -> Result<(Image, Option<OwnedFd>)> {
         let mut fence = MaybeUninit::uninit();
         let inner = construct_never_null(|res| {
             ffi::AImageReader_acquireNextImageAsync(self.as_ptr(), res, fence.as_mut_ptr())
@@ -220,7 +232,7 @@ impl ImageReader {
 
         Ok(match fence.assume_init() {
             -1 => (image, None),
-            fence => (image, Some(fence)),
+            fence => (image, Some(unsafe { OwnedFd::from_raw_fd(fence) })),
         })
     }
 
@@ -236,11 +248,15 @@ impl ImageReader {
         Ok(Some(Image { inner: res? }))
     }
 
+    /// Acquire the latest [`Image`] from the image reader's queue asynchronously, dropping older images.
+    ///
     /// # Safety
-    /// If the returned file descriptor is not `None`, it must be awaited before attempting to access the Image returned.
+    /// If the returned file descriptor is not [`None`], it must be awaited before attempting to
+    /// access the [`Image`] returned.
+    ///
     /// <https://developer.android.com/ndk/reference/group/media#aimagereader_acquirelatestimageasync>
     #[cfg(feature = "api-level-26")]
-    pub fn acquire_latest_image_async(&self) -> Result<(Image, Option<RawFd>)> {
+    pub fn acquire_latest_image_async(&self) -> Result<(Image, Option<OwnedFd>)> {
         let mut fence = MaybeUninit::uninit();
         let inner = construct_never_null(|res| unsafe {
             ffi::AImageReader_acquireLatestImageAsync(self.as_ptr(), res, fence.as_mut_ptr())
@@ -250,7 +266,7 @@ impl ImageReader {
 
         Ok(match unsafe { fence.assume_init() } {
             -1 => (image, None),
-            fence => (image, Some(fence)),
+            fence => (image, Some(unsafe { OwnedFd::from_raw_fd(fence) })),
         })
     }
 }
@@ -350,8 +366,8 @@ impl Image {
     }
 
     #[cfg(feature = "api-level-26")]
-    pub fn delete_async(self, release_fence_fd: RawFd) {
-        unsafe { ffi::AImage_deleteAsync(self.as_ptr(), release_fence_fd) };
+    pub fn delete_async(self, release_fence_fd: OwnedFd) {
+        unsafe { ffi::AImage_deleteAsync(self.as_ptr(), release_fence_fd.into_raw_fd()) };
         std::mem::forget(self);
     }
 }